pyproject.toml
[project]
name = "pixelart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"Pillow>=10.0.0",
	"webcolors>=1.13",
]
